strict check compile process

+++ include/Reducer.h
@@ -296,11 +296,11 @@ std::vector<std::string> Reducer<key, value>::Reduce(const std::vector<std::stri
     file_it->second.insert(std::make_pair(split_name, 1));
     master_it = master_map_.find(master_pair);
     file_it = master_it->second->files_.find(file_name);
-    if (file_it->second.size() == split_num) {
+    if (file_it->second.size() == static_cast<size_t>(split_num)) {
         master_it->second->count_++;
     }
 
-    if (master_it->second->count_ == master_it->second->files_.size()) {
+    if (static_cast<size_t>(master_it->second->count_) == master_it->second->files_.size()) {
         // 所有文件接收完毕,可以开始执行
         master_node->receive_all_.store(true);
         while (master_node->disk_merge_.load());
